Verify JWT Token Middleware & Handle Relationship User & Contact Schema and Protecting Routes

+++ controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require("../models/contactModel");
 
 // @desc Get all contacts
 // @route GET /api/contancts
-// @access public
+// @access private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 // @desc Create new contacts
 // @route POST /api/contancts
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res) => {
   console.log("The request body is :", req.body);
   const { name, email, phone } = req.body;
@@ -24,13 +24,14 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json(contact);
 });
 
 // @desc Get contacts
 // @route GET /api/contancts/:id
-// @access public
+// @access private
 
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -43,12 +44,18 @@ const getContact = asyncHandler(async (req, res) => {
 
 // @desc update contacts
 // @route PUT /api/contancts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Cotact Not Found");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      " User don't have permission to update other user contacts"
+    );
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -62,12 +69,18 @@ const updateContact = asyncHandler(async (req, res) => {
 
 // @desc Delete contacts
 // @route PUT /api/contancts/:id
-// @access public
+// @access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      " User don't have permission to update other user contacts"
+    );
   }
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json({ message: "Contact deleted successfully" });
